Rename `import_stock_trades_local` to `import_stock_trades_file`
